Tps/TP2/docker/app/requirements.txt
pandas
minio
pyarrow
s3fs
psycopg2-binary
